requirements.txt
openai>=1.30.0
anthropic>=0.25.0
google-generativeai>=0.5.0
datasets>=2.19.0
python-dotenv>=1.0.0
sentence-transformers>=2.7.0
pandas>=2.2.0
numpy>=1.26.0
matplotlib>=3.8.0
scikit-learn>=1.4.0